RecStrip 0.8: - Ersetzung von long durch int (Kompatibilität mit 32- & 64-bit Systemen)

+++ RecStrip.h
@@ -94,7 +94,7 @@ typedef struct
 
 typedef struct
 {
-  long                  NrBookmarks;
+  int                   NrBookmarks;
   dword                 Bookmarks[177];
   dword                 Resume;
 } TYPE_Bookmark_Info;
